+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Added
 - High speed trail behind players
 - Player squish effect 
+- Chronosphere OST
+- Airflow OST
 
 ### Fixed
 - Chat's input field unfocusing after scene changed
@@ -14,6 +16,8 @@
 - Overhauled landmine
 - Increased grenade explosion radius
 - Player color now displays in leaderboard
+- There is now 4 musics for each map
 
 ### Other
 - Player nickname and color now sets only when client connects
+- Music offset is now calculated by subtracting start date and current date